--- src/plugin/coreplugins/ExportImport/metadata.json
@@ -1,43 +1,0 @@
-{
-  "id": "ExportImport",
-  "name": "Export, Import and Update Library",
-  "version": "1.7.0",
-  "description": "Example of how to export, import and update a library from and to webgme.\n The active node (i.e. open node on the canvas) will be the starting point, expect when importing a project.",
-  "icon": {
-    "src": "",
-    "class": "glyphicon glyphicon-refresh"
-  },
-  "disableServerSideExecution": false,
-  "disableBrowserSideExecution": false,
-  "configStructure": [
-    {
-      "name": "type",
-      "displayName": "Type",
-      "description": "What library action should be taken?",
-      "value": "Export",
-      "valueType": "string",
-      "valueItems": [
-        "Export",
-        "ImportProject",
-        "ImportLibrary",
-        "UpdateLibrary"
-      ]
-    },
-    {
-      "name": "file",
-      "displayName": "Import file(s)",
-      "description": "WebGME library file and optionally exported assets.",
-      "value": "",
-      "valueType": "asset",
-      "readOnly": false
-    },
-    {
-      "name": "assets",
-      "displayName": "Export assets",
-      "description": "Export all encountered assets and return a zip package.",
-      "value": false,
-      "valueType": "boolean",
-      "readOnly": false
-    }
-  ]
-}
