+++ test/plugin/coreplugins/ImportV1/assets/assetsProject.json
