+++ test/plugin/coreplugins/ImportV1/assets/b.txt
